requirements.txt
# ── PyTorch (CUDA 12.4) ──────────────────────────────────────────────────────
# Install separately with:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124

# ── Core Deep Learning ────────────────────────────────────────────────────────
timm>=0.9.12              # ConvNeXt V2 pretrained models
Pillow>=10.0              # Image loading

# ── Data & Evaluation ────────────────────────────────────────────────────────
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3

# ── Visualization ────────────────────────────────────────────────────────────
matplotlib>=3.7
seaborn>=0.12

# ── Web UI ────────────────────────────────────────────────────────────────────
flask>=3.0.0
gunicorn>=21.2.0
opencv-python-headless>=4.8.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.65
thop>=0.1.1               # FLOPs counter for computational cost reporting
